Replace quic-go to fix CVE-2024-53259
There's a CVE in quic-go that's fixed in v0.48.2. We have an indirect
dependency on quic-go through CoreDNS. Instead of updating CoreDNS on a
release branch, replace the quic-go version.

Note that `go mod graph` still shows the older version, but it will not
actually be used. Graph reports the original module declarations, while
tidy and the build actually download the updated version.

    $ make LOCAL_BUILD=1 BUILD_UPX=false build
    $ go version -m bin/linux/amd64/lighthouse-coredns | grep quic-go
      dep  github.com/quic-go/quic-go  v0.42.0
      =>   github.com/quic-go/quic-go  v0.48.2

Signed-off-by: Daniel Farrell <dfarrell@redhat.com>